@@ -6,13 +6,14 @@
   "author": "Filip Szczesniak <fiszczu@gmail.com>",
   "license": "MIT",
   "dependencies": {
+    "@types/node": "^13.13.5",
     "antlr4": "^4.8.0",
     "cross-env": "^7.0.2",
-    "typescript": "^3.8.3"
+    "typescript": "3.9.3"
   },
   "scripts": {
-    "compile": "cross-env NODE_ENV=production node ./src/FiszczLangCompiler.js ./src/program.fs",
-    "compile:test": "cross-env NODE_ENV=production node ./src/FiszczLangCompiler.js ./src/test.fs",
+    "compile": "tsc && cross-env NODE_ENV=production node ./src/FiszczLangCompiler.js ./src/program.fs",
+    "compile:test": "tsc && cross-env NODE_ENV=production node ./src/FiszczLangCompiler.js ./src/test.fs",
     "genarate:test": "antlr4 ./src/FiszczLang.g4 & javac ./src/FiszczLang*.java",
     "test:gui": "grun FiszczLang program test.fs -gui",
     "generate": "antlr4 -Dlanguage=JavaScript -visitor ./src/FiszczLang.g4 -o ./src"
